server/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
supabase
cryptography
aioimaplib
passlib[bcrypt]
python-socketio[asyncio]
python-jose[cryptography]
python-dotenv